add skip to useHistory
